# Improve performance of embedded string allocation